Add an eeschema library ODBC back-end
The code needs rigorous testing and has heavy dependencies on RDBMS schema; in
particular, a stored procedures to assist in revisioning of the symbol data.
This schema info will be provided in a README file eventually, but that doesn't
exist yet.

This commit also contains minor updates to the myodbc plugin for pcbnew.